Added the ability to self-compare Outposts and Scouts
Before this implementation, if we had Outposts and Scouts with any overlap in instances (for instance, comparing a dataset of restaurants to itself), then the Outposts would count themselves. If self_comparison is set to True, then when analyzing Scouts, Outposts will not consider Scouts that are within about 5 feet (which we consider to be the same location as the Outpost).